Dualvnc

+++ tools/bin/vncstart.sh
@@ -26,7 +26,10 @@ if [ -d /etc/X11/fontpath.d ]; then
 fi
 
 # Start the VNC server
-if [ -x /usr/bin/Xvnc-realvnc ]; then
+RET=1 && (vnclicense -check >/dev/null 2>&1) && RET=$?
+if [[ -x /usr/bin/Xvnc-realvnc && -f /usr/lib/JARVICE/tools/etc/realvnc.key && $RET -eq 0 ]]; then
+    echo "Enabling RealVNC server for VNC service"
+    sudo cp -f /etc/NAE/help-real.html /etc/NAE/help.html
     mkdir -p ~/.vnc/config.d
     chmod 700 ~/.vnc/config.d
     cp -f /usr/lib/JARVICE/tools/etc/realvnc.key ~/.vnc/private.key
@@ -36,13 +39,14 @@ if [ -x /usr/bin/Xvnc-realvnc ]; then
 Password=$VNCPASSWD
 EOF
     touch ~/.vnc/config.d/.Xvnc-v5-marker
-    vncserver -geometry "$VNC_GEOMETRY" -StartUI=1 -EnableAutoUpdateChecks=0 \
+    /usr/bin/vncserver -geometry "$VNC_GEOMETRY" -StartUI=1 -EnableAutoUpdateChecks=0 \
         -ConnNotifyTimeout=0 \
         -AllowHttp=0 -Encryption PreferOn -Authentication VncAuth \
         -DisableAddNewClient -EnableRemotePrinting=0 -dpi 100 \
         -SecurityTypes RA2:256+,RA2,RA2ne,VeNCrypt,TLSVnc,VncAuth $FP :1
 else
-    vncserver -geometry "$VNC_GEOMETRY" \
+    # Default path to Tiger in /usr/local/bin
+    /usr/local/bin/vncserver -geometry "$VNC_GEOMETRY" \
         -rfbauth /etc/JARVICE/vncpasswd \
         -dpi 100 \
         -SecurityTypes=VeNCrypt,TLSVnc,VncAuth :1
